requirement.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
bitsandbytes
Pillow
datasets
peft
numpy
tqdm
matplotlib
opencv-python
easyocr
rembg
onnxruntime
